[FEAT] #12 - 식단 리서치 API

+++ server/src/main/kotlin/zzigmug/server/entity/Dish.kt
@@ -13,5 +13,5 @@ class Dish (
     var food: Food,
 
     @Column
-    var amount: Float,
+    var amount: Double,
 ): BaseEntity()
